remove hash so GA can't see it
One problem begets two solutions. I removed the hash from address bar as the page loads and save it for the load handler. This prevents GA from seeing it and also allows us to add it to the blob url that loads the wiki. If the access_token parameter is not present, then the complete hash string will carry through, allowing permalinks, permaviews, and safe mode. The hash is added to the blob URL before navigation, allowing it to work transparently.